+++ prod/streams/ksqldb/cleanup.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -e
+
+docker run --network nuvla-backend -v $(pwd):/nuvla -w /nuvla --rm -u root \
+        --entrypoint ./ksqldb-cleanup.sh confluentinc/cp-ksqldb-cli:6.0.0 \
+        ksqldb-server:8088